requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
colorama==0.4.4
cryptography==36.0.2
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
UserManager==0.5.3
Werkzeug==2.1.1
whitenoise==6.0.0
